+++ dataforge-io/src/commonMain/kotlin/space/kscience/dataforge/io/Envelope.kt
@@ -36,6 +36,8 @@ public interface Envelope {
 
 public class SimpleEnvelope(override val meta: Meta, override val data: Binary?) : Envelope
 
+public fun Envelope(meta: Meta, data: Binary?): Envelope = SimpleEnvelope(meta, data)
+
 /**
  * The purpose of the envelope
  *